refactor: remove `fsct-ipc` crate and migrate functionality
Eliminated the `fsct-ipc` crate and moved its JSON-RPC message types and framing helpers to `fsct-client`. Updated the driver and client code to directly integrate `serde`-based serialization/deserialization for `TimelineInfo` and `PlayerState`. Simplified imports and removed redundant code across the workspace.
