+++ clients/rust/src/rpc.rs
