+++ driver/Cargo.toml
@@ -27,7 +27,6 @@ serde = { version = "1.0.219", features = ["derive"] }
 tokio-util.workspace = true
 uuid.workspace = true
 fsct-client.workspace = true
-fsct-ipc.workspace = true
 serde_json.workspace = true
 futures.workspace = true
 nusb = { git = "https://github.com/HEM-RnD/nusb.git", tag = "v0.1.14-hem" }
@@ -59,6 +58,5 @@ async-stream = "0.3"
 
 [dev-dependencies]
 nix = { version = "0.29", default-features = false, features = ["signal", "poll", "fs"] }
-fsct-ipc.workspace = true
 serde_json.workspace = true
 tokio-util.workspace = true
