--- ipc/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "fsct-ipc"
-edition.workspace = true
-version.workspace = true
-authors.workspace = true
-license.workspace = true
-publish.workspace = true
-readme.workspace = true
-repository.workspace = true
-description = "JSON-RPC 2.0 message types and framing for FSCT IPC."
-
-[dependencies]
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
